Update and simplify v1 project description
Change the structure of the document to roughly reflect the abstract - start with the escape hatch because it's the most innovative part, and then go on to describe normal operation and instant transactions.
Defer scaling the escape hatch to v2 or later by linking a WIP doc that roughly describes the ideas.