Fix mint.js script.
It was generating the wrong input values for the circuit because it
was using Poseidon with the prime field of the bn254 curve. I just
forked the dependency and changes to the prime field of one of the Pasta
curves.

Note, there will be a significant challenge using poseidon over a cycle
of curves, since the hashes will change! This is a challenge for later.